eshop/src/main/java/com/epam/preprod/pavlov/servlet/LogoutServlet.java
package com.epam.preprod.pavlov.servlet;

import com.epam.preprod.pavlov.constant.PathConstants;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;
import java.io.IOException;

@WebServlet("/home/logout")
public class LogoutServlet extends AbstractServlet {
    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        HttpSession session = req.getSession();
        session.invalidate();
        resp.sendRedirect(req.getContextPath()+PathConstants.PRODUCT_SERVLET);

    }
}
